fix(stats): Round 3 hardening -- MC6 AR1 time-order, ER2/ER4 guards, ER6 dtype
Round 3 medium correctness/robustness items (stats-suite-review-gp.md):

- MC6 (gp.py, lme/_corrfit.py): ar1/car1 pair *consecutive within-group rows*;
  with time=None that silently assumes the rows are already in within-group time
  order. Warn (fit_corr_gls + gp_fit) when an order-dependent structure is fit
  without time=, and document the assumption in build_group_layout. cs is
  order-invariant and stays silent.
- ER2 (hgp.py): the nested model requires a GLOBALLY-numbered inner factor (each
  inner level nested within one outer level). Per-outer numbering (each outer's
  inner labels restarting at 0) passed the [0, L2) range check but aliased
  distinct subjects across outer groups -> mis-pool. Validate that each inner
  label pairs with a single outer label, else raise.
- ER4 (lme/reml.py): a 1-D random covariate z of shape (N,) is a single random
  slope; coerce it to the (N, r) contract so r = z.shape[-1] = 1 (a bare (N,)
  previously set r = N and misrouted into a deep IndexError); a malformed z
  raises a clear (N, r) error.
- ER6 (gp.py, hgp.py, inference/cluster.py, basis.py): close dtype leaks under
  the fp64 invariant -- promote an integer response Y to float in gp_fit/hgp_fit
  (an int Y coerced the whole fit to int); cluster_size_map counts in the
  canonical float (f64 under x64, not hardcoded f32 -- exact past 2**24); and
  mrf_smooth's adjacency promotes via issubdtype (not result_type(int, f32),
  which is f32 under JAX promotion), with re-eval in the stored design dtype.

Tests: ar1-without-time warns (cs / with-time silent); nested per-outer
numbering raises; 1-D z == (N, 1) z; integer-Y gp_fit/hgp_fit promote to float.
hgp + lme + lme_nested + lme_dispatch suites (69) green.

+++ src/nitrix/stats/inference/cluster.py
@@ -44,7 +44,10 @@ def cluster_size_map(
 
     Background voxels (label ``0``) map to ``0``.
     """
-    return _component_totals(labels, jnp.ones(labels.shape, dtype=jnp.float32))
+    # ER6: count in the canonical float (f64 under x64), not a hardcoded f32 --
+    # cluster sizes past 2**24 are not exactly representable in float32, eroding
+    # the TFCE / cluster-extent fp64 exactness.
+    return _component_totals(labels, jnp.ones(labels.shape, dtype=float))
 
 
 def cluster_mass_map(
